// apps/api/src/assets/dto/create-asset.dto.d.ts
export declare class CreateAssetDto {
    name: string;
    description?: string;
    serialNumber?: string;
    category?: string;
    type?: string;
    brand?: string;
    model?: string;
    size?: string;
    color?: string;
    condition?: string;
    vendor?: string;
    quantity?: number;
    licenseNumber?: string;
    complianceChecked?: boolean;
    status?: string;
    purchaseDate?: string;
    purchaseCost?: number;
    warrantyExpiration?: string;
    locationId?: string;
    assignedToId?: string;
    assignedDate?: string;
    expectedReturnDate?: string;
    notes?: string;
}